move cycle_detection in validate pkg
